+++ movie/src/test/kotlin/no/ecm/movie/TestBase.kt
@@ -3,8 +3,8 @@ package no.ecm.movie
 import io.restassured.RestAssured
 import io.restassured.RestAssured.*
 import io.restassured.http.ContentType
-import no.ecm.movie.response.GenreResponse
-import no.ecm.movie.response.MovieResponse
+import no.ecm.utils.response.GenreResponse
+import no.ecm.utils.response.MovieResponse
 import no.ecm.utils.dto.movie.GenreDto
 import no.ecm.utils.dto.movie.MovieDto
 import org.junit.After
