+++ utils/src/main/kotlin/no/ecm/utils/response/GenreResponse.kt
@@ -1,4 +1,4 @@
-package no.ecm.movie.response
+package no.ecm.utils.response
 
 import no.ecm.utils.dto.movie.GenreDto
 import no.ecm.utils.hal.PageDto
